fix: resolve Next.js route export type error for build-proof callback
